\name{SparseSignals-package}
\alias{SparseSignals-package}
\alias{SparseSignals}
\docType{package}
\title{Efficient algorithms for sparse signals}
\description{For piecewise constant signals with M or N segments
defined on 1, ..., L, (L >> M >= N), arithmetic operations can
be efficiently computed in O(M + N) rather than O(L) space and time.}
\details{
\tabular{ll}{Package: \tab SparseSignals\cr
Maintainer: \tab Toby Dylan Hocking <tdhock5@gmail.com>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 2014.3.11\cr
License: \tab GPL-3\cr
Title: \tab Efficient algorithms for sparse signals\cr
Suggests: \tab testthat\cr}
}
\author{Toby Dylan Hocking}

\keyword{ package }


